services/paddleocr-api/requirements.txt
fastapi>=0.115,<1.0
numpy>=2.1,<3.0
paddleocr==3.4.0
paddlepaddle==3.3.0
pillow>=11.0,<12.0
uvicorn[standard]>=0.34,<1.0
